// src/generated/prisma/enums.ts
export const PouchStatus = {
  CREATED: 'CREATED',
  IN_TRANSIT: 'IN_TRANSIT',
  RECEIVED: 'RECEIVED',
  CLOSED: 'CLOSED',
  LOST: 'LOST'
} as const

export type PouchStatus = (typeof PouchStatus)[keyof typeof PouchStatus]